Update t002.py
explicit casting of reward, returned by environment's step function, to float32 eliminates dqn.update()'s failure, which was saying: "Found dtype Double but expected Float"

+++ t002.py
@@ -247,7 +247,7 @@ class CForex(gym.Env):
 
         self.m_iLastAction=iAction        #    Close position
 
-        return copy.deepcopy(self.m_afObservation), fReward, done, oInfo
+        return copy.deepcopy(self.m_afObservation), np.float32(fReward), done, oInfo
 
     def render(self, mode: str = 'rgb_array'):
         """Renders the environment.
